Stability fixes (#75)
